ai-module/requirements.txt
fastapi
uvicorn
numpy
nltk
gensim
pinecone
python-dotenv
transformers
torch
langchain
langchain-core
langchain-google-genai
google-generativeai
langchain-huggingface
transformers
huggingface-hub
python-dotenv
numpy
scikit-learn

